pyproject.toml
[tool.poetry]
name = "specable"
version = "0.1.0"
description = "dictionaries to objects and vice versa"
authors = ["Marcel Langer <dev@marcel.science>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
nose = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
